Refactor stores and backend access
Now, stores only depend on a backend-technology independent interface.
Only current implementation  of this interface is a supabase backend. Technically, this can now be easily ported to another backend (e. g. PocketBase).
More importantly, it allows all the backend logic to be shared as library with other frontends (the sli.dev widget)